Split out transformer into additional files.
