+++ modules/angular2/src/transform/codegen.dart
@@ -3,71 +3,199 @@ library angular2.transformer;
 import 'package:analyzer/src/generated/ast.dart';
 import 'package:analyzer/src/generated/element.dart';
 import 'package:analyzer/src/generated/java_core.dart';
+import 'package:barback/barback.dart' show AssetId, TransformLogger;
+import 'package:dart_style/dart_style.dart';
+import 'package:path/path.dart' as path;
 
-String codegenClassTypeString(ClassElement el,
-                              Map<LibraryElement, String> libraryPrefixes) {
-  return '${_getPrefixDot(libraryPrefixes, el.library)}${el.name}';
-}
+import 'annotation_processor.dart';
 
-/// Creates the 'annotations' property for the Angular2 [registerType] call
-/// for [el].
-String codegenAnnotationsProp(ClassElement el,
-                             Map<LibraryElement, String> libraryPrefixes) {
-  var writer = new PrintStringWriter();
-  var visitor = new _AnnotationsTransformVisitor(writer, libraryPrefixes);
-  el.node.accept(visitor);
-  return writer.toString();
-}
+/// Base class that maintains codegen state.
+class Context {
+  final TransformLogger _logger;
+  /// Maps libraries to the import prefixes we will use in the newly
+  /// generated code.
+  final Map<LibraryElement, String> _libraryPrefixes;
 
-/// Creates the 'factory' property for the Angular2 [registerType] call
-/// for [ctor].
-String codegenFactoryProp(ConstructorElement ctor,
-                          Map<LibraryElement, String> libraryPrefixes) {
-  if (ctor.node == null) {
-    // This occurs when the class does not declare a constructor.
-    var prefix = _getPrefixDot(libraryPrefixes, ctor.type.element.library);
-    return '() => new ${prefix}${ctor.enclosingElement.displayName}()';
-  } else {
-    var writer = new PrintStringWriter();
-    var visitor = new _FactoryTransformVisitor(writer, libraryPrefixes);
-    ctor.node.accept(visitor);
-    return writer.toString();
+  DirectiveRegistry _directiveRegistry;
+  DirectiveRegistry get directiveRegistry => _directiveRegistry;
+
+  Context({TransformLogger logger})
+      : _logger = logger,
+        _libraryPrefixes = {} {
+    _directiveRegistry = new _DirectiveRegistryImpl(this);
+  }
+
+  void error(String errorString) {
+    if (_logger != null) {
+      _logger.error(errorString);
+    } else {
+      throw new Error(errorString);
+    }
+  }
+
+  /// If elements in [lib] should be prefixed in our generated code, returns
+  /// the appropriate prefix followed by a `.`. Future items from the same
+  /// library will use the same prefix.
+  /// If [lib] does not need a prefix, returns the empty string.
+  String _getPrefixDot(LibraryElement lib) {
+    var prefix = lib != null && !lib.isInSdk
+        ? _libraryPrefixes.putIfAbsent(lib, () => 'i${_libraryPrefixes.length}')
+        : null;
+    return prefix == null ? '' : '${prefix}.';
   }
 }
 
-/// Creates the 'parameters' property for the Angular2 [registerType] call
-/// for [ctor].
-String codegenParametersProp(ConstructorElement ctor,
-                             Map<LibraryElement, String> libraryPrefixes) {
-  if (ctor.node == null) {
-    // This occurs when the class does not declare a constructor.
-    return 'const [const []]';
+abstract class DirectiveRegistry {
+  // Adds [entry] to the `registerType` calls which will be generated.
+  void register(AnnotationMatch entry);
+}
+
+const _reflectorImport =
+    'import \'package:angular2/src/reflection/reflection.dart\' '
+    'show reflector;';
+
+/// Default implementation to map from [LibraryElement] to [AssetId]. This
+/// assumes that [el.source] has a getter called [assetId].
+AssetId _assetIdFromLibraryElement(LibraryElement el) {
+  return (el.source as dynamic).assetId;
+}
+
+String codegenEntryPoint(Context context,
+    {LibraryElement entryPoint, AssetId newEntryPoint}) {
+  // This must be called prior to [codegenImports] or the entry point
+  // library will not be imported.
+  var entryPointPrefix = context._getPrefixDot(entryPoint);
+  // TODO(jakemac): copyright and library declaration
+  var outBuffer = new StringBuffer(_reflectorImport);
+  _codegenImports(context, newEntryPoint, outBuffer);
+  outBuffer
+    ..write('main() {')
+    ..write(context.directiveRegistry.toString())
+    ..write('${entryPointPrefix}main();}');
+
+  return new DartFormatter().format(outBuffer.toString());
+}
+
+String _codegenImports(
+    Context context, AssetId newEntryPoint, StringBuffer buffer) {
+  context._libraryPrefixes.forEach((lib, prefix) {
+    buffer
+      ..write(_codegenImport(
+          context, _assetIdFromLibraryElement(lib), newEntryPoint))
+      ..writeln('as ${prefix};');
+  });
+}
+
+_codegenImport(Context context, AssetId libraryId, AssetId entryPoint) {
+  if (libraryId.path.startsWith('lib/')) {
+    var packagePath = libraryId.path.replaceFirst('lib/', '');
+    return "import 'package:${libraryId.package}/${packagePath}'";
+  } else if (libraryId.package != entryPoint.package) {
+    context._error("Can't import `${libraryId}` from `${entryPoint}`");
+  } else if (path.url.split(libraryId.path)[0] ==
+      path.url.split(entryPoint.path)[0]) {
+    var relativePath =
+        path.relative(libraryId.path, from: path.dirname(entryPoint.path));
+    return "import '${relativePath}'";
   } else {
-    var writer = new PrintStringWriter();
-    var visitor = new _ParameterTransformVisitor(writer, libraryPrefixes);
-    ctor.node.accept(visitor);
-    return writer.toString();
+    context._error("Can't import `${libraryId}` from `${entryPoint}`");
   }
 }
 
-String _getPrefixDot(Map<LibraryElement, String> libraryPrefixes,
-                     LibraryElement lib) {
-  var prefix = lib != null && !lib.isInSdk
-      ? libraryPrefixes.putIfAbsent(lib, () => 'i${libraryPrefixes.length}')
-      : null;
-  return prefix == null ? '' : '${prefix}.';
+class _DirectiveRegistryImpl implements DirectiveRegistry {
+  final Context _context;
+  final StringBuffer _buffer = new StringBuffer();
+
+  _DirectiveRegistryImpl(this._context);
+
+  @override
+  String toString() {
+    return _buffer.isEmpty ? '' : 'reflector${_buffer};';
+  }
+
+  // Adds [entry] to the `registerType` calls which will be generated.
+  void register(AnnotationMatch entry) {
+    var element = entry.element;
+    var annotation = entry.annotation;
+
+    if (annotation.element is! ConstructorElement) {
+      _context._error('Unsupported annotation type. '
+          'Only constructors are supported as Directives.');
+      return;
+    }
+    if (element is! ClassElement) {
+      _context._error('Directives can only be applied to classes.');
+      return;
+    }
+    if (element.node is! ClassDeclaration) {
+      _context._error('Unsupported annotation type. '
+          'Only class declarations are supported as Directives.');
+      return;
+    }
+    final ConstructorElement ctor = element.unnamedConstructor;
+    if (ctor == null) {
+      _context._error('No default constructor found for ${element.name}');
+      return;
+    }
+
+    _buffer.writeln('..registerType(${_codegenClassTypeString(element)}, {'
+        '"factory": ${_codegenFactoryProp(ctor)},'
+        '"parameters": ${_codegenParametersProp(ctor)},'
+        '"annotations": ${_codegenAnnotationsProp(element)}'
+        '})');
+  }
+
+  String _codegenClassTypeString(ClassElement el) {
+    return '${_context._getPrefixDot(el.library)}${el.name}';
+  }
+
+  /// Creates the 'annotations' property for the Angular2 [registerType] call
+  /// for [el].
+  String _codegenAnnotationsProp(ClassElement el) {
+    var writer = new PrintStringWriter();
+    var visitor = new _AnnotationsTransformVisitor(writer, _context);
+    el.node.accept(visitor);
+    return writer.toString();
+  }
+
+  /// Creates the 'factory' property for the Angular2 [registerType] call
+  /// for [ctor].
+  String _codegenFactoryProp(ConstructorElement ctor) {
+    if (ctor.node == null) {
+      // This occurs when the class does not declare a constructor.
+      var prefix = _context._getPrefixDot(ctor.type.element.library);
+      return '() => new ${prefix}${ctor.enclosingElement.displayName}()';
+    } else {
+      var writer = new PrintStringWriter();
+      var visitor = new _FactoryTransformVisitor(writer, _context);
+      ctor.node.accept(visitor);
+      return writer.toString();
+    }
+  }
+
+  /// Creates the 'parameters' property for the Angular2 [registerType] call
+  /// for [ctor].
+  String _codegenParametersProp(ConstructorElement ctor) {
+    if (ctor.node == null) {
+      // This occurs when the class does not declare a constructor.
+      return 'const [const []]';
+    } else {
+      var writer = new PrintStringWriter();
+      var visitor = new _ParameterTransformVisitor(writer, _context);
+      ctor.node.accept(visitor);
+      return writer.toString();
+    }
+  }
 }
 
 /// Visitor providing common methods for concrete implementations.
 abstract class _TransformVisitor extends ToSourceVisitor {
-
+  final Context _context;
   final PrintWriter _writer;
 
-  /// Map of [LibraryElement] to its associated prefix.
-  final Map<LibraryElement, String> _libraryPrefixes;
-
-  _TransformVisitor(PrintWriter writer, this._libraryPrefixes)
-  : this._writer = writer, super(writer);
+  _TransformVisitor(PrintWriter writer, this._context)
+      : this._writer = writer,
+        super(writer);
 
   /// Safely visit the given node.
   /// @param node the node to be visited
@@ -107,9 +235,9 @@ abstract class _TransformVisitor extends ToSourceVisitor {
   Object visitSimpleIdentifier(SimpleIdentifier node) {
     // Make sure the identifier is prefixed if necessary.
     if (node.bestElement is ClassElementImpl ||
-    node.bestElement is PropertyAccessorElement) {
+        node.bestElement is PropertyAccessorElement) {
       _writer
-        ..print(_getPrefixDot(_libraryPrefixes, node.bestElement.library))
+        ..print(_context._getPrefixDot(node.bestElement.library))
         ..print(node.token.lexeme);
     } else {
       return super.visitSimpleIdentifier(node);
@@ -123,8 +251,8 @@ class _CtorTransformVisitor extends _TransformVisitor {
   bool _withParameterTypes = true;
   bool _withParameterNames = true;
 
-  _CtorTransformVisitor(PrintWriter writer, libraryPrefixes)
-  : super(writer, libraryPrefixes);
+  _CtorTransformVisitor(PrintWriter writer, Context _context)
+      : super(writer, _context);
 
   /// If [_withParameterTypes] is true, this method outputs [node]'s type
   /// (appropriately prefixed based on [_libraryPrefixes]. If
@@ -132,7 +260,7 @@ class _CtorTransformVisitor extends _TransformVisitor {
   Object _visitNormalFormalParameter(NormalFormalParameter node) {
     if (_withParameterTypes) {
       var paramType = node.element.type;
-      var prefix = _getPrefixDot(_libraryPrefixes, paramType.element.library);
+      var prefix = _context._getPrefixDot(paramType.element.library);
       _writer.print('${prefix}${paramType.displayName}');
       if (_withParameterNames) {
         _visitNodeWithPrefix(' ', node.identifier);
@@ -151,8 +279,8 @@ class _CtorTransformVisitor extends _TransformVisitor {
   @override
   Object visitFieldFormalParameter(FieldFormalParameter node) {
     if (node.parameters != null) {
-      throw new Error('Parameters in ctor not supported '
-      '(${super.visitFormalParameterList(node)}');
+      _context.error('Parameters in ctor not supported '
+          '(${super.visitFormalParameterList(node)}');
     }
     return _visitNormalFormalParameter(node);
   }
@@ -183,8 +311,8 @@ class _CtorTransformVisitor extends _TransformVisitor {
 /// ToSourceVisitor designed to print 'parameters' values for Angular2's
 /// [registerType] calls.
 class _ParameterTransformVisitor extends _CtorTransformVisitor {
-  _ParameterTransformVisitor(PrintWriter writer, libraryPrefixes)
-  : super(writer, libraryPrefixes);
+  _ParameterTransformVisitor(PrintWriter writer, Context _context)
+      : super(writer, _context);
 
   @override
   Object visitConstructorDeclaration(ConstructorDeclaration node) {
@@ -213,8 +341,8 @@ class _ParameterTransformVisitor extends _CtorTransformVisitor {
 /// ToSourceVisitor designed to print 'factory' values for Angular2's
 /// [registerType] calls.
 class _FactoryTransformVisitor extends _CtorTransformVisitor {
-  _FactoryTransformVisitor(PrintWriter writer, libraryPrefixes)
-  : super(writer, libraryPrefixes);
+  _FactoryTransformVisitor(PrintWriter writer, Context _context)
+      : super(writer, _context);
 
   @override
   Object visitConstructorDeclaration(ConstructorDeclaration node) {
@@ -233,9 +361,8 @@ class _FactoryTransformVisitor extends _CtorTransformVisitor {
 /// ToSourceVisitor designed to print a [ClassDeclaration] node as a
 /// 'annotations' value for Angular2's [registerType] calls.
 class _AnnotationsTransformVisitor extends _TransformVisitor {
-
-  _AnnotationsTransformVisitor(PrintWriter writer, libraryPrefixes)
-  : super(writer, libraryPrefixes);
+  _AnnotationsTransformVisitor(PrintWriter writer, Context _context)
+      : super(writer, _context);
 
   @override
   Object visitClassDeclaration(ClassDeclaration node) {
