+++ modules/angular2/test/transform/ctor_with_default_value_files/expected/index.bootstrap.dart
@@ -1,14 +1,14 @@
 import 'package:angular2/src/reflection/reflection.dart' show reflector;
-import 'index.dart' as i0;
-import 'bar.dart' as i1;
-import 'package:angular2/src/core/annotations/annotations.dart' as i2;
+import 'bar.dart' as i0;
+import 'package:angular2/src/core/annotations/annotations.dart' as i1;
+import 'index.dart' as i2;
 
 main() {
   reflector
-    ..registerType(i1.Component, {
-      "factory": (dynamic c) => new i1.Component(c),
+    ..registerType(i0.Component, {
+      "factory": (dynamic c) => new i0.Component(c),
       "parameters": const [const [dynamic]],
-      "annotations": const [const i2.Directive(context: 'soup')]
+      "annotations": const [const i1.Directive(context: 'soup')]
     });
-  i0.main();
+  i2.main();
 }
