+++ modules/angular2/test/transform/html_entry_point_files/index.dart
@@ -2,5 +2,4 @@ library web_foo;
 
 import 'package:angular2/src/core/annotations/annotations.dart';
 
-void main() {
-}
+void main() {}
