+++ modules/angular2/test/transform/list_of_types_files/expected/index.bootstrap.dart
@@ -1,15 +1,15 @@
 import 'package:angular2/src/reflection/reflection.dart' show reflector;
-import 'index.dart' as i0;
-import 'bar.dart' as i1;
-import 'foo.dart' as i2;
-import 'package:angular2/src/core/annotations/annotations.dart' as i3;
+import 'bar.dart' as i0;
+import 'foo.dart' as i1;
+import 'package:angular2/src/core/annotations/annotations.dart' as i2;
+import 'index.dart' as i3;
 
 main() {
   reflector
-    ..registerType(i1.Component, {
-      "factory": (i2.MyContext c) => new i1.Component(c),
-      "parameters": const [const [i2.MyContext]],
-      "annotations": const [const i3.Directive(context: const [i2.MyContext])]
+    ..registerType(i0.Component, {
+      "factory": (i1.MyContext c) => new i0.Component(c),
+      "parameters": const [const [i1.MyContext]],
+      "annotations": const [const i2.Directive(context: const [i1.MyContext])]
     });
-  i0.main();
+  i3.main();
 }
